Extract waveform interaction model
